feat: add default configurations to toml file
