+++ .gitignore
@@ -22,6 +22,9 @@ MANIFEST
 locales/*.pot
 locales/**/*.mo
 
+logs/
+undo/
+
 # See http://help.github.com/ignore-files/ for more about ignoring files.
 
 # Compiled output
