+++ config.toml
@@ -2,3 +2,22 @@
 mode = false
 host = "0.0.0.0"
 port = 54321
+
+[keyring]
+keyring_token_name = 'TODOIST-TEMPLATE'
+
+[todoist]
+undo_folder = "undo"
+
+[log]
+cli_level = 20 # logging.INFO
+cli_format = "%(message)s"
+file_level = 40 # logging.ERROR
+file_format = "%(asctime)s %(levelname)-5s %(name)s (%(processName)s) %(message)s"
+file_folder = "logs"
+file_rolling_max_byte = 52428800
+file_rolling_count = 10
+
+[python]
+min = [3,9]
+max = [4,0]
