+++ lib/key_ring.py
@@ -1,5 +1,6 @@
 """Keyring functions"""
 
+import logging
 import keyring
 from lib.i18n import _
 
@@ -18,5 +19,16 @@ def setup(service_id):
     """Prompt user for token"""
     token = input(_("Please enter your API token: "))
     set_api_token(service_id, token)
+
+
+def get_keyring_api_token(service_id, prompt=False):
+    """Get API Token from keyring"""
+    api_token = get_api_token(service_id)
+    if api_token is None and prompt:
+        while not api_token:
+            logging.warning(_("Todoist API token not found for %s application."), service_id)
+            setup(service_id)
+            api_token = get_api_token(service_id)
+    return api_token
 
 # ~@:-]
