--- todoist_template_cli.py
@@ -1,154 +1,0 @@
-#!/usr/bin/env python
-
-"""Easily add tasks to Todoist with customizable YAML templates"""
-
-import os
-import sys
-import datetime
-import logging
-import argparse
-import lib.__version__ as version
-from lib.i18n import _
-from lib import cli
-
-
-def _parse_cmd_line():
-    parser = argparse.ArgumentParser(
-        prog="todoist_template.py",
-        usage='%(prog)s [options]',
-        description=_('Easily add tasks to Todoist with customizable YAML templates')
-    )
-
-    # positional arguments:
-    parser.add_argument(
-        "template",
-        nargs="?",  # a single value, which can be optional
-        type=argparse.FileType("r", encoding="utf8"),
-        default=sys.stdin)
-
-    # options
-    parser.add_argument(
-        "-D",
-        dest="placeholders",
-        type=cli.val_placeholder,  # can be a file or a comma separated list of key=value
-        metavar="KEY0=VAL0,KEY1=VAL1...",
-        default={},
-        help=_("the placeholder values replaced in template")
-    )
-
-    parser.add_argument(
-        "--id",
-        dest="service_id",
-        default=cli.DEFAULT_KEYRING_TOKEN_NAME,
-        help=_("keyring service name where store Todoist API Token")
-    )
-
-    command_group = parser.add_mutually_exclusive_group()
-    command_group.add_argument(
-        "-d",
-        "--debug",
-        dest="loglevel",
-        default=logging.INFO,
-        action="store_const",
-        const=logging.DEBUG,
-        help=_("more verbose output"),
-    )
-    command_group.add_argument(
-        "-q",
-        "--quiet",
-        dest="loglevel",
-        default=logging.INFO,
-        action="store_const",
-        const=logging.NOTSET,
-        help=_("suppress output"),
-    )
-
-    parser.add_argument(
-        "--dry-run",
-        dest="dry_run",
-        default=False,
-        action="store_true",
-        help=_("allows the %(prog)s command to run a trial without making \
-any changes on Todoist.com, this process has the same output as the real \
-execution except for new object IDs."),
-    )
-
-    parser.add_argument(
-        "-u",
-        "--update",
-        dest="is_update",
-        default=False,
-        action="store_true",
-        help=_("update task with the same name instead of adding a new one")
-    )
-
-    parser.add_argument(
-        "--token",
-        dest="token",
-        help=_("the Todoist authorization token")
-    )
-
-    parser.add_argument(
-        "--undo",
-        dest="undo",
-        type=argparse.FileType("rb"),
-        help=_("loads undo file and rollbacks all operations in it")
-    )
-
-    parser.add_argument(
-        "--version",
-        action="version",
-        version="%(prog)s " + version.__version__,
-        help=_("show program's version number and exit"))
-
-    return parser.parse_args()
-
-
-def main():
-    """Main function"""
-    args = _parse_cmd_line()
-
-    logging.basicConfig(level=args.loglevel, format="%(message)s")
-    logging.info(version.LOGO)
-
-    try:
-        cli.check_python_version()
-
-        api_token = args.token if args.token else cli.get_keyring_api_token(args.service_id, True)
-
-        if args.undo:
-            cli.undo(args.undo, api_token)
-        else:
-            with args.template as file:
-                logging.debug(_("open file %s"), file)
-
-                script_folder = os.path.dirname(os.path.realpath(sys.argv[0]))
-                if not args.dry_run:
-                    now = datetime.datetime.now().strftime('%Y%m%d%H%M%S')
-                    undofile = os.path.join(
-                        script_folder,
-                        f"{os.path.basename(file.name)}-{now}.undo"
-                    )
-                else:
-                    undofile = None
-
-                cli.run_batch_template(
-                    cli.get_template_content(file),
-                    args.placeholders,
-                    api_token,
-                    args.dry_run,
-                    args.is_update,
-                    undofile
-                )
-
-        return 0
-
-    except Exception as exc:  # pylint: disable=broad-except
-        logging.error(exc, exc_info=logging.getLogger().isEnabledFor(logging.DEBUG))
-        return 1
-
-
-if __name__ == "__main__":
-    sys.exit(main())
-
-# ~@:-]
